examples/requirements_minio.txt
celery
redis
minio
boto3
python-decouple
pandas
fastparquet
s3fs